refactor: merged set end
