--- internal/handler/leasing_history_handler_set_end_timestamp.go
@@ -1,38 +1,0 @@
-package handler
-
-import (
-	"github.com/PitiNarak/condormhub-backend/pkg/apperror"
-	"github.com/gofiber/fiber/v2"
-	"github.com/google/uuid"
-)
-
-// SetEndTimestamp godoc
-// @Summary Delete a leasing history
-// @Description Delete a leasing history in the database
-// @Tags history
-// @Security Bearer
-// @Produce json
-// @Param id path string true "LeasingHistoryId"
-// @Success 204 "Set end timestamp successfully"
-// @Failure 400 {object} dto.ErrorResponse "Incorrect UUID format"
-// @Failure 401 {object} dto.ErrorResponse "your request is unauthorized"
-// @Failure 404 {object} dto.ErrorResponse "leasing history not found"
-// @Failure 500 {object} dto.ErrorResponse "Can not parse UUID or Failed to update leasing history"
-// @Router /history/{id} [patch]
-func (h *LeasingHistoryHandler) SetEndTimestamp(c *fiber.Ctx) error {
-	id := c.Params("id")
-
-	if err := uuid.Validate(id); err != nil {
-		return apperror.BadRequestError(err, "Incorrect UUID format")
-	}
-
-	leasingHistoryID, err := uuid.Parse(id)
-	if err != nil {
-		return apperror.InternalServerError(err, "Can not parse UUID")
-	}
-	err = h.service.SetEndTimestamp(leasingHistoryID)
-	if err != nil {
-		return err
-	}
-	return c.SendStatus(fiber.StatusNoContent)
-}
